My first DS project
Well, it was hard because I learnt only theory but now I had to remember all the functions, features and so on. Of course, I had to google a lot of things. I think it's okay since it's my first time. I have to get the skills and I hope won't have to google too much in the future.
     Well, what did I do? I tried a lot of things. Didn't know where and what to start with. Looked through some tutorials and books and got some clue of the data. Filled in some missing values. Have to finish cleaning.